tweak(web): rework admin raw workflow
