+++ web/src/components/AdminFilter.tsx
@@ -1,14 +1,23 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import { Switch } from "@/components/base/Switch";
 import { Select } from "@/components/base/Select";
 import { TextInput } from "@/components/base/TextInput";
 import { Button } from "@/components/base/Button";
-import { FilterOptions } from "@/components/LianeMapAdmin";
 import { IndexedRawTrip } from "@/api";
 
 interface FilterProps {
   callback: (filterOptions: FilterOptions) => void,
+  load: () => void,
   rawTrips: IndexedRawTrip[]
+}
+
+export interface FilterOptions {
+  chosenUser?: string;
+  chosenTrip?: number;
+  displayBackground: boolean;
+  displayForeground: boolean;
+  distanceBetweenPoints?: number;
+  timeBetweenPoints?: number;
 }
 
 function extractIndex(rawTrips: IndexedRawTrip[]) {
@@ -30,53 +39,40 @@ function extractUsers(rawTrips: IndexedRawTrip[]) {
   return users;
 }
 
-export function AdminFilter({ callback, rawTrips }: FilterProps) {
-  const [displayRawTrips, setDisplayRawTrips] = useState(true);
-  const [allUsers, setAllUsers] = useState(true);
-  const [chosenUser, setChosenUser] = useState<string>();
-  const [allTrips, setAllTrips] = useState(true);
-  const [chosenTrip, setChosenTrip] = useState<number>();
+export function AdminFilter({ callback, load, rawTrips }: FilterProps) {
+  const [chosenUser, setSelectedUser] = useState<string>();
+  const [chosenTrip, setSelectedTrip] = useState<number>();
   const [displayBackground, setDisplayBackground] = useState(true);
   const [displayForeground, setDisplayForeground] = useState(true);
   const [distanceBetweenPoints, setDistanceBetweenPoints] = useState<number>();
   const [timeBetweenPoints, setTimeBetweenPoints] = useState<number>();
 
-  function cb() {
+  // Update dynamically
+
+  useEffect(() => {
     callback({
-      displayRawTrips,
-      allUsers,
       chosenUser,
       displayBackground,
       displayForeground,
       distanceBetweenPoints,
       timeBetweenPoints,
-      allTrips,
       chosenTrip
     });
-  }
-
-  function selectTripController(id) {
-    if ((typeof id === "string") || (id === 0)) {
-      setAllTrips(true);
-    } else {
-      setAllTrips(false);
-    }
-    setChosenTrip(id);
-  }
-
-  function selectUserController(user) {
-    if (user === "Tous les utilisateurs") {
-      setAllUsers(true);
-    } else {
-      setAllUsers(false);
-    }
-    setChosenUser(user);
-  }
+  }, [chosenUser, displayBackground, displayForeground, distanceBetweenPoints, timeBetweenPoints, chosenTrip]);
 
   return (
     <div className="absolute top-0 right-0 z-10 overflow-auto">
       <div className="bg-white w-96 shadow-xl bg-opacity-60 rounded-lg grid grid-cols-2 p-6 gap-2 m-6">
-        <Switch label="Données brutes" value={displayRawTrips} onChange={setDisplayRawTrips} color="yellow" />
+        <span>Charger les trajets</span>
+
+        <Button
+          color="orange"
+          className="mt-4 col-span-2"
+          label="Charger"
+          onClick={load}
+        />
+
+        <span>Filtrer les trajets</span>
 
         <Select
           className="col-span-2"
@@ -84,7 +80,7 @@ export function AdminFilter({ callback, rawTrips }: FilterProps) {
           options={Array.from(extractUsers(rawTrips))}
           value={chosenUser}
           render={(id) => id}
-          onChange={(id) => selectUserController(id)}
+          onChange={setSelectedUser}
           placeholder="Aucun"
         />
 
@@ -94,9 +90,7 @@ export function AdminFilter({ callback, rawTrips }: FilterProps) {
           options={Array.from(extractIndex(rawTrips))}
           value={chosenTrip}
           render={(id) => (!(id === 0) ? id : null)}
-          onChange={(id) => {
-            selectTripController(id);
-          }}
+          onChange={setSelectedTrip}
         />
 
         <TextInput
@@ -117,13 +111,6 @@ export function AdminFilter({ callback, rawTrips }: FilterProps) {
 
         <Switch label="Données background" value={displayBackground} onChange={setDisplayBackground} color="yellow" />
         <Switch label="Données foreground" value={displayForeground} onChange={setDisplayForeground} color="yellow" />
-
-        <Button
-          color="orange"
-          className="mt-4 col-span-2"
-          label="Valider"
-          onClick={() => { cb(); }}
-        />
       </div>
     </div>
   );
